requirement.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
Django==4.2
django-crum==0.7.9
django-debug-toolbar==4.0.0
django-filter==24.3
django-ratelimit==4.1.0
djangorestframework==3.15.2
idna==3.8
lxml==5.3.0
lxml_html_clean==0.2.2
Mako==1.3.5
Markdown==3.7
MarkupSafe==2.1.5
nh3==0.2.18
requests==2.32.3
setuptools==74.1.2
sqlparse==0.5.1
urllib3==2.2.3
wrapt==1.16.0
